requirements.txt
authlib
python-dotenv
flask
requests
ndjson
python-chess